+++ src/app/core/mock-data/ccc-api-response.ts
