+++ src/app/core/services/corporate-credit-card-expense.service.spec.ts
@@ -5,7 +5,7 @@ import { ApiService } from './api.service';
 import { AuthService } from './auth.service';
 import { CorporateCreditCardExpenseService } from './corporate-credit-card-expense.service';
 import { DataTransformService } from './data-transform.service';
-import { apiCardV2Transactions } from '../test-data/corporate-credit-card-expense.spec.data';
+import { apiCardV2Transactions } from '../mock-data/ccc-api-response';
 import { expectedECccResponse } from '../mock-data/corporate-card-expense-unflattened.data';
 import { uniqueCardsParam } from '../mock-data/unique-cards.data';
 import { cardAggregateStatParam } from '../mock-data/card-aggregate-stat.data';
@@ -94,6 +94,7 @@ describe('CorporateCreditCardExpenseService', () => {
           group_id: testID,
         },
       });
+      expect(dataTransformService.unflatten).toHaveBeenCalledTimes(1);
       done();
     });
   });
